+++ ashley-extensions/src/com/roaringcatgames/kitten2d/ashley/K2EntityTweenAccessor.java
@@ -2,6 +2,7 @@ package com.roaringcatgames.kitten2d.ashley;
 
 import aurelienribon.tweenengine.TweenAccessor;
 import com.badlogic.ashley.core.Entity;
+import com.badlogic.gdx.Gdx;
 import com.roaringcatgames.kitten2d.ashley.components.*;
 
 /**
@@ -60,6 +61,7 @@ public class K2EntityTweenAccessor implements TweenAccessor<Entity> {
                 if(K2ComponentMappers.transform.has(entity)){
                     TransformComponent tc = K2ComponentMappers.transform.get(entity);
                     returnValues[0] = tc.position.z;
+                    Gdx.app.log("TweenAccessor", "Z: " + returnValues[0]);
                     result = 1;
                 }
                 break;
@@ -161,6 +163,7 @@ public class K2EntityTweenAccessor implements TweenAccessor<Entity> {
                 break;
             case POSITION_Z:
                 if(K2ComponentMappers.transform.has(entity)){
+                    Gdx.app.log("TweenAccessor", "New Z: " + newValues[0]);
                     TransformComponent tc = K2ComponentMappers.transform.get(entity);
                     tc.position.set(tc.position.x, tc.position.y, newValues[0]);
                 }