+++ ashley-extensions/src/com/roaringcatgames/kitten2d/ashley/systems/ZComparator.java
@@ -1,24 +1,22 @@
 package com.roaringcatgames.kitten2d.ashley.systems;
 
-import com.badlogic.ashley.core.ComponentMapper;
 import com.badlogic.ashley.core.Entity;
-import com.roaringcatgames.kitten2d.ashley.components.TransformComponent;
+import com.roaringcatgames.kitten2d.ashley.K2ComponentMappers;
 
 import java.util.Comparator;
 
 /**
- * Created by barry on 12/8/15 @ 10:22 PM.
+ * A comparator to Order by Z index
  */
 public class ZComparator implements Comparator<Entity> {
-    private ComponentMapper<TransformComponent> transformM;
 
     public ZComparator(){
-        transformM = ComponentMapper.getFor(TransformComponent.class);
+
     }
 
     @Override
     public int compare(Entity entityA, Entity entityB) {
-        return (int) Math.signum(transformM.get(entityB).position.z -
-                transformM.get(entityA).position.z);
+        return (int) Math.signum(K2ComponentMappers.transform.get(entityB).position.z -
+                K2ComponentMappers.transform.get(entityA).position.z);
     }
 }